+++ Source/Properties/AssemblyInfo.cs
@@ -30,6 +30,6 @@ using System.Runtime.InteropServices;
 // You can specify all the values or you can default the Build and Revision Numbers 
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion ("1.14.*")]
-[assembly: AssemblyInformationalVersion ("1.14 for KSP 1.4+")]
-[assembly: KSPAssembly ("KIS", 1, 14)]
+[assembly: AssemblyVersion ("1.15.*")]
+[assembly: AssemblyInformationalVersion ("1.15 for KSP 1.5+")]
+[assembly: KSPAssembly ("KIS", 1, 15)]