+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.15 (pre-release):
+# 1.15 (October 27th, 2018):
 * [Change] KSP 1.5.* compatibility.
 * [Fix #268] Wrong size detection on the MH expansion parts. READ #273!!!
 * [Fix #269] Equip items and remove helmet actions don't work.